requirements.txt
streamlit
streamlit-webrtc
python-dotenv
openai>=1.55.0
pydub
av
numpy
streamlit-mic-recorder